Ready first_common_ancestor function
