+++ cracking_practices/get_node_path/get_node_path.py
@@ -1,10 +1,3 @@
-# First Common Ancestor 4.8
-
-# Design an algorithm and write the code to find the first common ancestor of two nodes in a binary tree.
-# Avoid storing additional nodes in the data structure.
-# _NOTE This is not necessarily a binary search tree.
-
-
 class Node:
     def __init__(self, value):
         self.value = value
@@ -167,81 +160,4 @@ def get_node_path(tree, node):
 
     traverse(tree.root, node)
     return return_stack
-
-
-def first_common_ancestor(tree, node_a, node_b):
-    common_ancestor = None
-
-    if not tree or not node_a or not node_b:
-        return common_ancestor
-
-    path_node_a = get_node_path(tree, node_a)
-    path_node_b = get_node_path(tree, node_b)
-
-    for i in range(0, len(path_node_a)):
-        if path_node_a[i] in path_node_b:
-            return path_node_a[i]
-
-    return common_ancestor
-
-
-
-
-if __name__ == "__main__":
-    BST = BinarySearchTree()
-    # level 0
-    BST.add(100)
-    # level 1
-    BST.add(50)
-    BST.add(150)
-    # level 2
-    BST.add(25)
-    BST.add(75)
-    BST.add(125)
-    BST.add(175)
-    # level 3
-    BST.add(15)
-    BST.add(40)
-    BST.add(60)
-    BST.add(90)
-    BST.add(115)
-    BST.add(140)
-    BST.add(160)
-    BST.add(190)
-    # level 4
-    BST.add(5)
-    BST.add(20)
-    BST.add(30)
-    BST.add(45)
-    BST.add(55)
-    BST.add(65)
-    BST.add(80)
-    BST.add(95)
-    BST.add(110)
-    BST.add(120)
-    BST.add(130)
-    BST.add(145)
-    BST.add(151)
-    BST.add(165)
-    BST.add(180)
-    BST.add(200)
-
-    # Node: 20
-    node_a = BST.root.left.left.left.right
-
-    print(node_a.value)
-    print(get_node_path(BST, node_a))
-
-    # Node 90
-    node_b = BST.root.left.right.right
-    print(node_b.value)
-    print(get_node_path(BST, node_b))
-
-    print(first_common_ancestor(BST,node_a, node_b ))
-
-    # # Node 145
-    # node_b = BST.root.right.left.right.right
-    # print(node_b.value)
-    # print(get_node_path(BST, node_b))
-
